+++ activity/brightpay/nextjs/changelog.md
@@ -4642,3 +4642,14 @@
 
 ---
 
+# Brightpay Development - 2025-02-26
+
+**Commit #123**
+
+- Repository: brightpay/brightclinic
+- Type: healthcare-platform
+- Language: nextjs
+- Month: 2/2025
+
+---
+